Fix bug passing `region` to `subregion` in cache
 * Also add unit test for `InternationalAddress`
   json codec.
