+++ src/main/scala/com/socrata/geocoders/Geocoder.scala
@@ -33,55 +33,3 @@ trait OptionalGeocoder {
    */
   def geocode(addresses: Seq[Option[InternationalAddress]]): Seq[Option[LatLon]]
 }
-
-case class InternationalAddress(address: Option[String],
-                                locality: Option[String],
-                                subregion: Option[String],
-                                region: Option[String],
-                                postalCode: Option[String],
-                                country: String)
-
-object InternationalAddress {
-  @deprecated(message = "Confusing syntax; use create(.) instead.", since = "2.0.2")
-  def apply(address: Option[String],
-            locality: Option[String],
-            subregion: Option[String],
-            region: Option[String],
-            postalCode: Option[String],
-            country: Option[String]): Option[InternationalAddress] = {
-    create(address, locality, subregion, region, postalCode, country)
-  }
-
-  // filter out empty strings to None and default country to US
-  def create(address: Option[String],
-            locality: Option[String],
-            subregion: Option[String],
-            region: Option[String],
-            postalCode: Option[String],
-            country: Option[String]): Option[InternationalAddress] = {
-    if (Seq(address, locality, subregion, region, postalCode, country).forall(_.isEmpty)) None
-    else Some(new InternationalAddress(
-      address.filter(_.nonEmpty),
-      locality.filter(_.nonEmpty),
-      subregion.filter(_.nonEmpty),
-      region.filter(_.nonEmpty),
-      postalCode.filter(_.nonEmpty),
-      country.filter(_.nonEmpty).getOrElse("US")))
-  }
-}
-
-@deprecated(message = "Unused; use InternationalAddress.create(.) instead.", since = "2.0.2")
-object USAddress {
-  def apply(address: Option[String],
-            city: Option[String],
-            state: Option[String],
-            zip: Option[String]): Option[InternationalAddress] = {
-    InternationalAddress.create(address, city, None, state, zip, None)
-  }
-}
-
-case class LatLon(lat: Double, lon: Double)
-
-object LatLon {
-  implicit val llCodec = AutomaticJsonCodecBuilder[LatLon]
-}
